Improve formatting of displayed values
For each displayed metric in a row of an output table, use the
actual metric values, the difference between the different values
in the row, and the supplied units to come up with an appropriate
number of significant figures.

For seconds and bytes, pick between us/ms/s and B/KiB/MiB based on
the values, not based upon the raw suppplied units.